:pencil2: 타임캡슐 id 타입 변경
- Int -> Long

+++ core/navigation/src/main/java/com/idiotfrogs/navigation/Routes.kt
@@ -17,7 +17,7 @@ sealed interface Routes: NavKey {
     @Serializable
     data class Friend(val id: Int) : Routes
     @Serializable
-    data class Detail(val id: Int) : Routes
+    data class Detail(val id: Long) : Routes
     @Serializable
     data class Message(val id: Int) : Routes
 }
